+++ uliweb/template_files/project/wsgi_handler.py
@@ -4,5 +4,5 @@ path = os.path.dirname(os.path.abspath(__file__))
 if path not in sys.path:
     sys.path.insert(0, path)
 
-from uliweb.manage import make_application
-application = make_application(project_dir=path)
+from uliweb.manage import make_simple_application
+application = make_simple_application(project_dir=path)