+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Uliweb Change Log
 * Add `MIME_TYPES` section to upload app, but it'll only effect with uliweb application, but not web server
 * Improve call command, enable call modules ouside apps directory, and add project path to sys.path
 * Fix orm PICKLE update bug, use deepcopy to save old_value.
+* Add tornado server support.
 
 0.1.7 Version
 -----------------